requirements.txt
aiohttp==3.9.1
discord.py==2.3.2
pandas==2.1.4
protobuf==4.25.2
pytz==2023.3.post1
